Add redirect_model configuration option.
If 'redirect_model' is specified in the the 'jasper' section of trytond's
configuration file, jasper will call the method 'send_report' of the supplied
model instead of directly sending the report to the client, which is its default
behaviour.

So 'send_report' can be used to send the report directly to a given printer,
save it in a spool, etc.

It's been specifically designed for the new 'printer' module.